Get phenotype log

Get phenotype log

getPhenotypeLog(cohortIds = NULL)

Arguments

cohortIds

IDs of cohorts to extraction from the library.

Value

Returns a table with one row per cohort definitions with log information such as its release cycle. Example, this function gives us insight on when a cohort definition was added/updated/deprecated by the OHDSI PhenotypeLibrary.

A tibble.

Examples

getPhenotypeLog(cohortIds = c(1, 2))
#> # A tibble: 1 x 78
#>   cohortId cohortName         cohortNameAtlas cohortNameFormatted cohortNameLong
#>      <dbl> <chr>              <chr>           <chr>               <chr>         
#> 1        2 COVID-19 diagnosi~ [W] COVID-19 d~ COVID-19 diagnosis~ COVID-19 diag~
#> # i 73 more variables: librarian <chr>, status <chr>, addedVersion <chr>,
#> #   logicDescription <chr>, hashTag <chr>, isCirceJson <dbl>,
#> #   contributors <chr>, contributorOrcIds <chr>,
#> #   contributorOrganizations <chr>, peerReviewers <chr>,
#> #   peerReviewerOrcIds <dbl>, recommendedReferentConceptIds <chr>,
#> #   ohdsiForumPost <chr>, createdDate <date>, modifiedDate <date>,
#> #   lastModifiedBy <dbl>, replaces <dbl>, isReferenceCohort <dbl>, ...